requirements.txt
fastapi>=0.128
uvicorn>=0.40
python-dotenv>=1.2
httpx>=0.28
pyjwt>=2.11

# Workers
arq>=0.27
redis>=5.3,<6

